+++ src/main/java/org/rookit/dm/play/similarity/PlaylistComparator.java
@@ -27,8 +27,8 @@ public class PlaylistComparator extends AbstractPlayableComparator<StaticPlaylis
 	}
 
 	@Override
-	protected Map<String, Integer> createTopMap(StaticPlaylist element1, StaticPlaylist element2) {
-		final Map<String, Integer> scores = super.createTopMap(element1, element2);
+	protected Map<String, Double> createTopMap(StaticPlaylist element1, StaticPlaylist element2) {
+		final Map<String, Double> scores = super.createTopMap(element1, element2);
 		scores.put(NAME, compareStringIgnoreCase(element1.getName(), element2.getName()));
 		scores.put(TRACKS, reverseIntersect(element1.getTracks(), element2.getTracks()));
 		return scores;
